+++ CHANGELOG.md
@@ -2,13 +2,21 @@
 
 ## UNRELEASED
 
+## 3.1.0-M6 (November 16, 2018)
+
 ### FEATURES
 
+* Support GCE virtual private networks (VPC) ([GH-80](https://github.com/ystia/yorc/issues/80))
 * Support Kubernetes Jobs. ([GH-86](https://github.com/ystia/yorc/issues/86))
+
+### ENHANCEMENTS
+
+* Allow user to provide an already existing namespace to use when creating Kubernetes resources ([GH-76](https://github.com/ystia/yorc/issues/76))
 
 ### BUG FIXES
 
 * Generate unique names for GCP resources ([GH-177](https://github.com/ystia/yorc/issues/177))
+* Need a HOST public_ip_address attribute on Hosts Pool compute nodes ([GH-199](https://github.com/ystia/yorc/issues/199))
 
 ## 3.1.0-M5 (October 26, 2018)